feat: improve design with new content
