+++ components/partials/resources/ResourceCards.vue
@@ -7,7 +7,7 @@
       :key="`${item.title}-${i}`"
       class="light:bg-light-surface text-center dark:bg-dark-surface rounded p-6 light:hover:bg-gray-200 dark:hover:bg-gray-800 cursor-pointer transition-colors duration-300 ease-linear"
     >
-      <NuxtLink :to="`/with/${item.name}`">
+      <NuxtLink :to="`/resources/integrations/with/${item.name}`">
         <img
           v-if="item.logo"
           :src="item.logo"
